// client/src/store/mutations/types/appTypes.ts
export enum AppMutationType {
  SHOW_LOADING = "SHOW_LOADING",
  HIDE_LOADING = "HIDE_LOADING",
  SET_ERROR = "SET_ERROR",
  CLEAR_ERROR = "CLEAR_ERROR",
  SET_EDITING = "SET_EDITING",
  CANCEL_EDITING = "CANCEL_EDITING",
  SET_CREATING = "SET_CREATING",
  CANCEL_CREATING = "CANCEL_CREATING",
  SET_HOME = "SET_HOME",
  SET_CLONING = "SET_CLONING",
  CANCEL_CLONING = "CANCEL_CLONING",
}